refactor impl & add unittest
