+++ paddle/fluid/operators/activation_op.cc
@@ -219,6 +219,13 @@ $$out = \\frac{1}{\\sqrt{x}}$$
 
 )DOC";
 
+UNUSED constexpr char AbsDoc[] = R"DOC(	
+Abs Operator.	
+
+$$out = |x|$$	
+
+)DOC";
+
 UNUSED constexpr char CeilDoc[] = R"DOC(
 Ceil Operator. Computes ceil of x element-wise.
 
@@ -707,6 +714,7 @@ REGISTER_ACTIVATION_OP_MAKER(Tanh, TanhDoc);
 REGISTER_ACTIVATION_OP_MAKER(TanhShrink, TanhShrinkDoc);
 REGISTER_ACTIVATION_OP_MAKER(Sqrt, SqrtDoc);
 REGISTER_ACTIVATION_OP_MAKER(Rsqrt, RsqrtDoc);
+REGISTER_ACTIVATION_OP_MAKER(Abs, AbsDoc);
 REGISTER_ACTIVATION_OP_MAKER(Ceil, CeilDoc);
 REGISTER_ACTIVATION_OP_MAKER(Floor, FloorDoc);
 REGISTER_ACTIVATION_OP_MAKER(Cos, CosDoc);
@@ -782,6 +790,120 @@ class ActivationOpDoubleGrad2 : public framework::OperatorWithKernel {
   framework::OpKernelType GetExpectedKernelType(
       const framework::ExecutionContext& ctx) const override {
     return GetKernelType(ctx, *this, "DDX");
+  }
+};
+
+class AbsGradOp : public framework::OperatorWithKernel {
+ public:
+  using framework::OperatorWithKernel::OperatorWithKernel;
+
+  void InferShape(framework::InferShapeContext* ctx) const override {
+    auto out_grad_name = framework::GradVarName("Out");
+    ctx->ShareDim(out_grad_name, framework::GradVarName("X"));
+    ctx->ShareLoD(out_grad_name, framework::GradVarName("X"));
+  }
+
+ protected:
+  framework::OpKernelType GetExpectedKernelType(
+      const framework::ExecutionContext& ctx) const override {
+    return GetKernelType(ctx, *this, "X");
+  }
+
+  framework::OpKernelType GetKernelTypeForVar(
+      const std::string& var_name, const framework::Tensor& tensor,
+      const framework::OpKernelType& expected_kernel_type) const {
+    return framework::OpKernelType(tensor.type(), tensor.place(),
+                                   tensor.layout());
+  }
+};
+
+template <ActBwdOpFwdDeps kDepValue, typename T>
+class AbsGradOpMaker : public framework::SingleGradOpMaker<T> {
+ public:
+  using framework::SingleGradOpMaker<T>::SingleGradOpMaker;
+
+ protected:
+  void Apply(GradOpPtr<T> op) const override {
+    op->SetType(this->ForwardOpType() + "_grad");
+    op->SetInput("X", this->Input("X"));
+    op->SetInput(framework::GradVarName("Out"), this->OutputGrad("Out"));
+    op->SetOutput(framework::GradVarName("X"), this->InputGrad("X"));
+    op->SetAttrMap(this->Attrs());
+
+    if ((static_cast<int>(kDepValue) &
+         static_cast<int>(ActBwdOpFwdDeps::kDepX)) ||
+        FLAGS_use_mkldnn ||
+        (op->HasAttr("use_mkldnn") &&
+         BOOST_GET_CONST(bool, op->GetAttr("use_mkldnn")))) {
+      op->SetInput("X", this->Input("X"));
+    }
+
+    if (static_cast<int>(kDepValue) &
+        static_cast<int>(ActBwdOpFwdDeps::kDepOut)) {
+      op->SetInput("Out", this->Output("Out"));
+    }
+  }
+};
+
+template <ActBwdOpFwdDeps kDepValue>
+class AbsDoubleGradOp : public framework::OperatorWithKernel {
+ public:
+  using framework::OperatorWithKernel::OperatorWithKernel;
+
+  void InferShape(framework::InferShapeContext* ctx) const override {
+    if (static_cast<int>(kDepValue) & static_cast<int>(kDepX)) {
+      if (ctx->HasOutput("DX")) {
+        ctx->ShareDim("X", "DX");
+        ctx->ShareLoD("X", "DX");
+      }
+      if (ctx->HasOutput("DDOut")) {
+        ctx->ShareDim("X", "DDOut");
+        ctx->ShareLoD("X", "DDOut");
+      }
+    }
+    if (static_cast<int>(kDepValue) & static_cast<int>(kDepOut)) {
+      if (ctx->HasOutput("DOut")) {
+        ctx->ShareDim("Out", "DOut");
+        ctx->ShareLoD("Out", "DOut");
+      }
+      if (ctx->HasOutput("DDOut")) {
+        ctx->ShareDim("Out", "DDOut");
+        ctx->ShareLoD("Out", "DDOut");
+      }
+    }
+  }
+
+ protected:
+  framework::OpKernelType GetExpectedKernelType(
+      const framework::ExecutionContext& ctx) const override {
+    return GetKernelType(ctx, *this, "DDX");
+  }
+
+  framework::OpKernelType GetKernelTypeForVar(
+      const std::string& var_name, const framework::Tensor& tensor,
+      const framework::OpKernelType& expected_kernel_type) const {
+    return framework::OpKernelType(tensor.type(), tensor.place(),
+                                   tensor.layout());
+  }
+};
+
+// AbsGrad: dx=dy if x >=0 else -dy
+// AbsDoubleGrad: ddy = ddx if x >=0 else -ddx
+template <typename T>
+class AbsDoubleGradMaker : public ::paddle::framework::SingleGradOpMaker<T> {
+ public:
+  using ::paddle::framework::SingleGradOpMaker<T>::SingleGradOpMaker;
+
+ protected:
+  void Apply(GradOpPtr<T> op) const override {
+    op->SetType("abs_grad_grad");
+    // input1: x
+    op->SetInput("X", this->Input("X"));
+    // input2: ddx
+    op->SetInput("DDX", this->OutputGrad(framework::GradVarName("X")));
+    op->SetAttrMap(this->Attrs());
+    // output: ddy
+    op->SetOutput("DDOut", this->InputGrad(framework::GradVarName("Out")));
   }
 };
 
@@ -1292,6 +1414,67 @@ REGISTER_OP_CPU_KERNEL(
                               ops::ExpGradFunctor<int>>,
     ops::ActivationGradKernel<paddle::platform::CPUDeviceContext,
                               ops::ExpGradFunctor<int64_t>>);
+/* ========================================================================== */
+
+/* ==========================   abs register  ============================ */
+REGISTER_OPERATOR(
+    abs, ops::ActivationOp, ops::AbsOpMaker, ops::ActivationOpInferVarType,
+    ops::AbsGradOpMaker<ops::AbsGradFunctor<float>::FwdDeps(),
+                        paddle::framework::OpDesc>,
+    ops::AbsGradOpMaker<ops::AbsGradFunctor<float>::FwdDeps(),
+                        paddle::imperative::OpBase>,
+    std::conditional<ops::CanInplaceAct<ops::AbsGradFunctor<float>>(),
+                     ops::ActFwdInplaceInferer, void>::type);
+REGISTER_OPERATOR(abs_grad, ops::AbsGradOp, ops::ActivationGradOpInplaceInferer,
+                  ops::AbsDoubleGradMaker<paddle::framework::OpDesc>,
+                  ops::AbsDoubleGradMaker<paddle::imperative::OpBase>);
+REGISTER_OPERATOR(
+    abs_grad_grad,
+    ops::AbsDoubleGradOp<ops::AbsGradGradFunctor<float>::FwdDeps()>,
+    ops::ActivationDoubleGradOpInplaceInferer);
+
+REGISTER_OP_CPU_KERNEL(
+    abs, ops::ActivationKernel<paddle::platform::CPUDeviceContext,
+                               ops::AbsFunctor<float>>,
+    ops::ActivationKernel<paddle::platform::CPUDeviceContext,
+                          ops::AbsFunctor<double>>,
+    ops::ActivationKernel<paddle::platform::CPUDeviceContext,
+                          ops::AbsFunctor<int>>,
+    ops::ActivationKernel<paddle::platform::CPUDeviceContext,
+                          ops::AbsFunctor<int64_t>>,
+    ops::ActivationKernel<paddle::platform::CPUDeviceContext,
+                          ops::AbsFunctor<plat::complex64>>,
+    ops::ActivationKernel<paddle::platform::CPUDeviceContext,
+                          ops::AbsFunctor<plat::complex128>>);
+REGISTER_OP_CPU_KERNEL(
+    abs_grad, ops::ActivationGradKernel<paddle::platform::CPUDeviceContext,
+                                        ops::AbsGradFunctor<float>>,
+    ops::ActivationGradKernel<paddle::platform::CPUDeviceContext,
+                              ops::AbsGradFunctor<double>>,
+    ops::ActivationGradKernel<paddle::platform::CPUDeviceContext,
+                              ops::AbsGradFunctor<int>>,
+    ops::ActivationGradKernel<paddle::platform::CPUDeviceContext,
+                              ops::AbsGradFunctor<int64_t>>,
+    ops::ActivationGradKernel<paddle::platform::CPUDeviceContext,
+                              ops::AbsGradFunctor<plat::complex64>>,
+    ops::ActivationGradKernel<paddle::platform::CPUDeviceContext,
+                              ops::AbsGradFunctor<plat::complex128>>);
+REGISTER_OP_CPU_KERNEL(
+    abs_grad_grad,
+    ops::ActivationDoubleGradKernel<plat::CPUDeviceContext,
+                                    ops::AbsGradGradFunctor<float>>,
+    ops::ActivationDoubleGradKernel<plat::CPUDeviceContext,
+                                    ops::AbsGradGradFunctor<double>>,
+    ops::ActivationDoubleGradKernel<plat::CPUDeviceContext,
+                                    ops::AbsGradGradFunctor<plat::float16>>,
+    ops::ActivationDoubleGradKernel<plat::CPUDeviceContext,
+                                    ops::AbsGradGradFunctor<int>>,
+    ops::ActivationDoubleGradKernel<plat::CPUDeviceContext,
+                                    ops::AbsGradGradFunctor<int64_t>>,
+    ops::ActivationDoubleGradKernel<paddle::platform::CPUDeviceContext,
+                                    ops::AbsGradGradFunctor<plat::complex64>>,
+    ops::ActivationDoubleGradKernel<paddle::platform::CPUDeviceContext,
+                                    ops::AbsGradGradFunctor<plat::complex128>>);
 /* ========================================================================== */
 
 /* ==========================  Log register ==================================*/
