+++ paddle/fluid/operators/activation_op.cu
@@ -174,6 +174,51 @@ REGISTER_OP_CUDA_KERNEL(
                               ops::ExpGradFunctor<plat::float16>>);
 /* ========================================================================== */
 
+/* ==========================   abs register  ============================ */
+REGISTER_OP_CUDA_KERNEL(
+    abs, ops::ActivationKernel<plat::CUDADeviceContext, ops::AbsFunctor<float>>,
+    ops::ActivationKernel<plat::CUDADeviceContext, ops::AbsFunctor<double>>,
+    ops::ActivationKernel<plat::CUDADeviceContext, ops::AbsFunctor<int>>,
+    ops::ActivationKernel<plat::CUDADeviceContext, ops::AbsFunctor<int64_t>>,
+    ops::ActivationKernel<plat::CUDADeviceContext,
+                          ops::AbsFunctor<plat::float16>>,
+    ops::ActivationKernel<plat::CUDADeviceContext,
+                          ops::AbsFunctor<plat::complex64>>,
+    ops::ActivationKernel<plat::CUDADeviceContext,
+                          ops::AbsFunctor<plat::complex128>>);
+REGISTER_OP_CUDA_KERNEL(
+    abs_grad, ops::ActivationGradKernel<plat::CUDADeviceContext,
+                                        ops::AbsGradFunctor<float>>,
+    ops::ActivationGradKernel<plat::CUDADeviceContext,
+                              ops::AbsGradFunctor<double>>,
+    ops::ActivationGradKernel<plat::CUDADeviceContext,
+                              ops::AbsGradFunctor<int>>,
+    ops::ActivationGradKernel<plat::CUDADeviceContext,
+                              ops::AbsGradFunctor<int64_t>>,
+    ops::ActivationGradKernel<plat::CUDADeviceContext,
+                              ops::AbsGradFunctor<plat::float16>>,
+    ops::ActivationGradKernel<plat::CUDADeviceContext,
+                              ops::AbsGradFunctor<plat::complex64>>,
+    ops::ActivationGradKernel<plat::CUDADeviceContext,
+                              ops::AbsGradFunctor<plat::complex128>>);
+REGISTER_OP_CUDA_KERNEL(
+    abs_grad_grad,
+    ops::ActivationDoubleGradKernel<paddle::platform::CUDADeviceContext,
+                                    ops::AbsGradGradFunctor<float>>,
+    ops::ActivationDoubleGradKernel<paddle::platform::CUDADeviceContext,
+                                    ops::AbsGradGradFunctor<double>>,
+    ops::ActivationDoubleGradKernel<plat::CUDADeviceContext,
+                                    ops::AbsGradGradFunctor<plat::float16>>,
+    ops::ActivationDoubleGradKernel<paddle::platform::CUDADeviceContext,
+                                    ops::AbsGradGradFunctor<int>>,
+    ops::ActivationDoubleGradKernel<paddle::platform::CUDADeviceContext,
+                                    ops::AbsGradGradFunctor<int64_t>>,
+    ops::ActivationDoubleGradKernel<paddle::platform::CUDADeviceContext,
+                                    ops::AbsGradGradFunctor<plat::complex64>>,
+    ops::ActivationDoubleGradKernel<paddle::platform::CUDADeviceContext,
+                                    ops::AbsGradGradFunctor<plat::complex128>>);
+/* ========================================================================== */
+
 /* ==========================  Log register ==================================*/
 REGISTER_ACTIVATION_CUDA_KERNEL(log, Log, LogFunctor, LogGradFunctor);
